feat: add std::hash<tr::shared_string>
It makes shared_string usable as an unordered container key and
retires bespoke hashing in the GTK client.

hash_combine moves from net.h to utils.h because gtk cannot include
net.h, which is libtransmission-only.

+++ gtk/Torrent.h
@@ -100,7 +100,7 @@ public:
     Speed get_speed_up() const;
     tr_torrent& get_underlying() const;
     Storage get_total_size() const;
-    unsigned int get_trackers() const;
+    size_t get_trackers() const;
 
     Glib::RefPtr<Gio::Icon> get_icon() const;
     Glib::ustring get_short_status_text() const;
